pyproject.toml
[project]
name = "birbal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.127.0",
    "langchain>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-text-splitters>=1.1.0",
    "ollama>=0.6.1",
    "orgparse>=0.4.20251020",
    "pandas>=2.3.3",
    "psycopg>=3.3.2",
    "python-dotenv>=1.2.1",
]
[project.scripts]
birbal-server = "birbal.cli:run_server"
